Separate resource jar construction from compilation
This adds a new action to Java (and Android/Java proto) targets with
resources that add any resources to the output jar separately from compilation.
The main advantage of this approach is that changes to resources no longer
force the target to be recompiled.

It also makes actions that create resource jars but do no compilation easier to
set up, since they no longer pull in JavaBuilder and JavaTargetAttributes.

PiperOrigin-RevId: 152181755