Add a new CachedLocalSpawnRunner to implement remote caching + local exec
The new class wraps an existing SpawnRunner and adds remote caching. Ideally,
the wrapped runner should be local and sandboxed, but this is not currently
enforced. The new class is not hooked up to anything yet.

The added test indicates that the RemoteActionCache interface is still more
complex than necessary - in particular, we should merge downloadAllResults and
downloadBlobs (for stdout/stderr) into a single method, and also change the
upload to a single combined method in a similar way instead of two calls. Doing
so allows the RemoteActionCache implementation more leeway in how it wants to
implement these, potentially improving parallelism and performance.

One step towards #1413.

PiperOrigin-RevId: 152245644