Add a class providing a view of a NestedSet
When serializing a large number of related nested sets needs to be serialized
in an efficient way, it is necessary to access the internal structure of a nested
set to efficiently deduplicate shared content. Add a new class that provides such
a view on a nested set.

Note: part of this change (in particular, the addition of the NestedSetView class)
was accidentally committed as part of 83217139df4b09ccedba776b5412dd956ccd0481.

Change-Id: I03660a228a66bbd6d3df2d3e78e8349be2d55f41
PiperOrigin-RevId: 152362816