Do not trample the PersistentMap journal
This fixes https://github.com/bazelbuild/bazel/issues/2660. Basically,
if we elect to keep the journal during PersistentMap.save(), we
shouldn't stomp over it the next time save() is called.

In writeJournal(), we now check if the journal file exists, and open
it in append mode if it does. Alternatively, we could simply not close
(and thus forget about) the journal in save(), but that would leak the
journal file handle if save() was never called with keepJournal()
returning false.

Change-Id: Id00732f161c8b5a082a6c109aee115591ace2ea7
PiperOrigin-RevId: 152480978