Make depset.to_list() return a non-frozen list
Previously it returned a frozen list. Using frozen mutables as part of the API specification is an unnecessary complication and should be avoided.

RELNOTES[INC]: The return type of depset.to_list() is now a list rather than a frozen list. (Modifying the list has no effect on the depset.)
PiperOrigin-RevId: 152504943