Support multidex in Bazel.
Previously multidex did not work in Bazel because it required a dependency on
zip, which we do not ship with Bazel. Then the
--experimental_android_use_singlejar_for_multidex flag was added to use a
different tool that is included with Bazel for multidex. This change sets that
flag to be the default.

Fixes https://github.com/bazelbuild/bazel/issues/1936

RELNOTES: android_binary multidex should now work without additional flags.
PiperOrigin-RevId: 152712301