Automated g4 rollback of commit b0ac3b41700cdc575c36ff2de191bc9ba2b6abf0.
*** Reason for rollback ***

kokoro failures were not flakes :(

*** Original change description ***

Fix escaping of '@' in javabuilder command lines

"@" is used to denote params files, so when it appears in labels it is
escaped as "@@". The second "@" should be removed during flag parsing
to avoid e.g. having it show up in strict deps errors.

See bazelbuild/buildtools#83

PiperOrigin-RevId: 153127997