Canonicalize-flags can now show the canonical invocation policy.
Canonicalize-flags will provide the expanded and filtered version of the policy passed to it as a command argument if also passed the command arg --canonicalize_policy. This is the version that would be actually applied to the command line in another command.

In this mode, it will not also print out the interpretation of the flags passed after "--", so to get both outputs, it will have to be run twice. This was to keep the output clean and easy to parse.

RELNOTES: new option for bazel canonicalize-flags, --canonicalize_policy
PiperOrigin-RevId: 153215518