Vendor jarjar instead of using a binary version of it.
It appears that jarjar is not actually used except for renaming classes
in the coverage collecting Java test runner as verified by

`find . -name BUILD | xargs grep jarjar`

Yak shaving for #2820.

This is necessary because the current version of jarjar does not work
with Java 8 classes because it embeds a version of ASM that doesn't
support that yet.

Change-Id: I6ac59b84bbbc1e85fe8e7f4f4876b98fc6129df0