When using experimental UI, only to a reset terminal if curses were used
When unregistering an event handler, that uses colors, we need to be sure
to set the terminal in a normal state after unregistering that event handler.
The ExperimentalEventHandler does use curses, however only if options allow
to do so (otherwise, curses will be filtered out by the underlying terminal).
So, we cannot conclude the use of color from the fact that the
ExperimentalEventHandler was in use. Hence when releasing it, only reset the
the terminal, if we did use color and avoid the unnecessary ^[[0m; otherwise.

RELNOTES: None.
PiperOrigin-RevId: 154177220