Bazel client, Windows: impl. server->IsStillAlive
Implement proper aliveness check for the Bazel
server.

This allows detecting a server death that occurs
after the Java process started but before it
started the gRPC server.

On other platforms we open an anonymous pipe, let
the server inherit the writing end, close the
writing end in the client, and attempt to read
non-blockingly in order to check if the server is
still alive.

This approach fails on Windows because anonymous
pipes are always blocking. Named pipes support
asynchronous access but it's much simpler to just
check if the process exited. GetProcessTimes looks
like a reliable way to do so, and that's what we
use on Windows now.

Fixes https://github.com/bazelbuild/bazel/issues/2817

Change-Id: Ic24577d8440eb0c8188c852e2501ce1e254ba9fd
PiperOrigin-RevId: 154283585