BEP: Report exit code in BuildCompleteEvent
We decided to use the exit code as opposed to a status enum as bazel
reports the build status as an exit code internally and the number of exit
codes is dynamic. That is, a bazel module might add additional exit codes
and thus we need to support reporting these.

The overall_success field is now redundant and has thus been removed. A
build was successfull iff ExitCode.code equals zero.

Change-Id: I268db972b09d983f6cd07df34e3a384efb607358
PiperOrigin-RevId: 154413356