Allow bypassing the automated detection of jobs when --jobs=auto.
This is useful for the Blaze case where builds are cloud-backed.  In that
case, we need to use a very large number of jobs to handle the remote
tasks, as each task is handled by a blocking thread controlled by jobs.

We were previously doing this using an invocation policy, but the UI was
extremely confusing: "blaze help build" would claim that the default for
--jobs was "auto" when, internally, "auto" was being converted to the
hardcoded number pretty much silently.  However, explicitly specifying
--jobs=auto on the command line yielded different behaviors.

With this change, --jobs=auto works for both Bazel and Blaze and does
sensible things in each case without the user having to know.

RELNOTES: None.
PiperOrigin-RevId: 154418572