Fix bazel_toolchain_test
This time for real. This rolls back ea55d71799e4, my previous attempt to fix
the test. What happened:

Before d637f6c5c636:
The BUILD file was referencing the data as a directory, and the sandbox
implementation was symlinking the entire directory (not individual files).
As a result, the chmod call was succeeding, but *it was modifying the source
tree*!

After d637f6c5c636, before ea55d71799e4:
The BUILD file was using a glob to enumerate individual files, and the
individual symlinks in the sandbox are in a read-only file system. The chmod
call was failing.

After ea55d71799e4:
Without the chmod call, the test passed because the files were still marked
as executable *in the source tree* due to the pre-d637f6c5c636 code. However,
it failed in a clean checkout.

Solution:
Reinstate the chmod call, but make sure all the files are copied into a
temporary *empty* directory.

Fixes #2891.

PiperOrigin-RevId: 154436636