BuildOptions cleanup:
1) Remove obsolete originalOptions field. This was originally added to
   support "parent" transitions, which supported config1 -> config2 -> config1
   transitions by having config2 store config1's options.

   The purpose of this feature was to support LIPO (which has a DATA -> TARGET
   transition). But https://github.com/bazelbuild/bazel/commit/a09c3968ccd8d0b3a5526a95877c629ea90d5945
   makes this unnecessary.

2) Support the "disable actions" feature of the LIPO context collector configuration.
   Putting this in BuildOptions make this dynamic config-compatible.

This change intentionally doesn't add disableOptions to BuildOptions.equals() or
BuildOptions.hashCode(). It'd be great to do that. But that has semantic consequences.
And we've run into really tricky bugs in the past with dynamic configurations and
BuildOptions.equals / BuildConfiguration.equals. So it's best to experiment with
that in its own change.

PiperOrigin-RevId: 154999718