Rename some --incompatible_* flags to be more specific
This is possibly a nit, but we don't want to reuse flag names in the future, so it's a good idea to include what the flag *does* in the name as opposed to just the feature it affects, in case the same feature is changed multiple times. Updated javadoc for incompatible change system to say so.

This rename is ok because these flags have only been submitted over the past couple days.

RELNOTES: None
PiperOrigin-RevId: 155371363